add uploaded_document_service
