+++ app/services/bgs/people_service.rb
@@ -1,13 +1,8 @@
 # frozen_string_literal: true
 
 module BGS
-  class PeopleService
-    include SentryLogging
+  class PeopleService < BaseService
     class VaFileNumberNotFound < StandardError; end
-
-    def initialize(current_user)
-      @current_user = current_user
-    end
 
     def find_person_by_participant_id
       response = service.people.find_person_by_ptcpnt_id(@current_user.participant_id)
@@ -24,13 +19,7 @@ module BGS
     private
 
     def report_no_va_file_user(e)
-      log_exception_to_sentry(
-        e,
-        {
-          icn: @current_user.icn
-        },
-        { team: 'vfs-ebenefits' }
-      )
+      report_error(e)
 
       PersonalInformationLog.create(
         error_class: e,
@@ -42,15 +31,6 @@ module BGS
             participant_id: @current_user.participant_id
           }
         }
-      )
-    end
-
-    def service
-      external_key = @current_user.common_name || @current_user.email
-
-      @service ||= BGS::Services.new(
-        external_uid: @current_user.icn,
-        external_key: external_key
       )
     end
   end
