+++ app/services/bgs/uploaded_document_service.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module BGS
+  class UploadedDocumentService < BaseService
+    def get_documents
+      @service.uploaded_document.find_by_participant_id(@user.participant_id)
+    end
+  end
+end
